Default _DEFAULT_SOURCE macro when compiling
This tells the GNU compiler (gcc) to include the functions defined in
endian(3). This is not required but it prevents errors when trying to
use specific compiler like c99.

This makes the project dependent on the GNU compiler, which is not
ideal. Later, these functions should be defined locally so the C99
standard can be used.