Add testing capabilities
Adds rules for building tests to the Makefile. Tests go in the tests/
directory and are normal C source files. The final test binary gets
built from tests/main.c, so all code needs to be included in that file.

The test binary can be built using 'make test' and run using './test'.