Reduce output when building
Instead of printing the whole command line when invoking the compiler,
the output has been shortened to "CC	%filename%", where %filename% is
the name of the source file being compiled. Finally, when all the files
are being linked, "LINK	%bin%" is printed, where %bin% is the name of
the resulting binary.

Also, a 'print-flags' rule has been added to the Makefile to print the
*FLAGS variables since they are no longer printed when building.